+++ CHANGELOG.md
@@ -3,13 +3,16 @@ These are selected quick notes for developers. For all changes, see the
 
 8.x-3.3:
 
+* Configuration: sp_cert_folder has been removed. sp_private_key,
+  sp_x509_certificate, (new) sp_new_certificate, idp_certs and
+  idp_cert_encryption can now hold values with a 'file:' and 'key:' prefix
+  (followed by respectively an absolute filename and an entity ID of a 'Key'
+  entity, instead of the full contents of a key/certificate).
+
 * Configuration: idp_x509_certificate, idp_x509_certificate_multi and
   idp_cert_type have been removed. idp_certs and idp_cert_encryption have been
   added (with the idp_x509_certificate value moving to idp_certs and
   idp_x509_certificate_multi moving to either idp_certs or idp_cert_encryption).
-
-* Configuration: sp_cert_folder has been removed. sp_x509_certificate and
-  sp_private_key can now hold values with a 'file:' prefix.
 
 * SamlService::$samlAuth was changed into an array. (This is not considered
   part of the interface. SamlService::getSamlAuth(), which should be used for